+++ go.mod
@@ -2,4 +2,7 @@ module github.com/Ranadia/overengineered-calculator
 
 go 1.13
 
-require github.com/gorilla/mux v1.7.3
+require (
+	cloud.google.com/go/firestore v1.1.1
+	github.com/gorilla/mux v1.7.3
+)
