+++ model/calculation.go
@@ -1,8 +1,8 @@
 package model
 
 type Calculation struct {
-	typeOfCalculation string
-	firstNumber       float64
-	secondNumber      float64
-	result            float64
+	TypeOfCalculation string
+	FirstNumber       float64
+	SecondNumber      float64
+	Result            float64
 }
